Connect stomp api to routes
