+++ server/routes/arRoutes.js
@@ -3,7 +3,7 @@ const arController = require('../controllers/arController');
 
 const router = express.Router();
 
-router.get('/', arController.tourView);
-router.get('/d3', arController.d3Tests);
+router.get('/', arController.getApiData, arController.tourView);
+router.get('/d3', arController.getApiData, arController.d3Tests);
 
 module.exports = router;
