+++ server/views/arViews/d3.ejs
@@ -3,7 +3,10 @@
 
 	<h1>Homepage</h1>
 	<p><%= message %></p>
-  <p></p>
+  <pre><%= h.dump(data) %></pre>
+
+
+
   <svg width="960" height="500"></svg>
 
   <script src="https://d3js.org/d3.v4.min.js"></script>
